Edit

+++ src/iTechArt.Shook.Repositories/Stores/UserStore.cs
@@ -1,5 +1,4 @@
-﻿using iTechArt.Common;
-using iTechArt.Shook.DomainModel.Models;
+﻿using iTechArt.Shook.DomainModel.Models;
 using iTechArt.Shook.Repositories.UnitsOfWorks;
 using System;
 using System.Threading;
@@ -20,12 +19,19 @@ namespace iTechArt.Shook.Repositories.Stores
             _uow = uow;
         }
 
+        public void Dispose()
+        {
+            _uow.Dispose();
+        }
 
         public async Task<IdentityResult> CreateAsync(User user,
             CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
-            
+            if (user == null)
+            {
+                throw new ArgumentNullException();
+            }
             await _uow.UserRepository.CreateAsync(user);
             await _uow.SaveChangesAsync();
 
@@ -36,22 +42,27 @@ namespace iTechArt.Shook.Repositories.Stores
         {
             cancellationToken.ThrowIfCancellationRequested();
 
+            if (user == null)
+            {
+                throw new ArgumentNullException();
+            }
+
             _uow.UserRepository.Delete(user);
             await _uow.SaveChangesAsync();
 
             return IdentityResult.Success;
         }
 
-        public void Dispose()
-        {
-            _uow.Dispose();
-        }
+        
 
         public async Task<User> FindByIdAsync(string userId, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
 
-            int.TryParse(userId, out var id);
+            if (!int.TryParse(userId, out var id))
+            {
+                throw new InvalidCastException();
+            }
 
             var user = await _uow.UserRepository.GetByIdAsync(id);
 
@@ -61,31 +72,47 @@ namespace iTechArt.Shook.Repositories.Stores
         public async Task<User> FindByNameAsync(string userName, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
-
+            
             return await _uow.UserRepository.FirstOrDefaultAsync(u => u.UserName == userName);
         }
 
         public Task<string> GetNormalizedUserNameAsync(User user, CancellationToken cancellationToken)
         {
-            cancellationToken.ThrowIfCancellationRequested();
             throw new NotImplementedException();
         }
 
         public Task<string> GetUserIdAsync(User user, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
+
+            if (user == null)
+            {
+                throw new ArgumentNullException();
+            }
+
             return Task.FromResult(user.Id.ToString());
         }
 
         public Task<string> GetUserNameAsync(User user, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
+            
+            if (user == null)
+            {
+                throw new ArgumentNullException();
+            }
+
             return Task.FromResult(user.UserName);
         }
 
         public async Task<IdentityResult> UpdateAsync(User user, CancellationToken cancellationToken)
         {
             cancellationToken.ThrowIfCancellationRequested();
+
+            if (user == null)
+            {
+                throw new ArgumentNullException();
+            }
 
             _uow.UserRepository.Update(user);
             await _uow.SaveChangesAsync();
@@ -95,13 +122,11 @@ namespace iTechArt.Shook.Repositories.Stores
 
         public Task SetNormalizedUserNameAsync(User user, string normalizedName, CancellationToken cancellationToken)
         {
-            cancellationToken.ThrowIfCancellationRequested();
             return Task.FromResult<object>(null);
         }
 
         public Task SetUserNameAsync(User user, string userName, CancellationToken cancellationToken)
         {
-            cancellationToken.ThrowIfCancellationRequested();
             throw new NotImplementedException();
         }
     }
